update quiz create

+++ database/migrations/2020_07_06_081734_create_quizzes_table.php
@@ -14,9 +14,9 @@ class CreateQuizzesTable extends Migration
     public function up()
     {
         Schema::create('quizzes', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name');
-            $table->unsignedInteger('category_id');
+            $table->unsignedBigInteger('category_id');
             $table->tinyInteger('duration');
             $table->integer('question_count');
             $table->tinyInteger('published')->default(0);
